events: update event metadata timestamps and relay status handling

+++ crates/events/src/post/models.rs
@@ -13,7 +13,7 @@ pub struct RadrootsPostEventIndex {
 pub struct RadrootsPostEventMetadata {
     pub id: String,
     pub author: String,
-    pub published_at: u32,
+    pub published_at: u64,
     pub kind: u32,
     pub post: RadrootsPost,
 }
